requirements.txt
fastapi
uvicorn
python-multipart
openai
pydantic
pandas
python-docx
openpyxl
